Changed Window Object setup in effort to simplify
Tried to simplify overall in the hopes that makes the codemore readable and easy to debug. Removed distinction between visual and manual mode because it's not needed in the long-run. Removed drawing functions for now, we will refer to the past commits and add them back in as needed. Removed the older window measurements defintiion because we are not using them for now, and in the future the full set of window measurement defintions would actually be more complicated than what you have noted